POLIO-1070: refactor API to get shapes for all OU
