+++ plugins/polio/serializers.py
@@ -1266,121 +1266,3 @@ class ConfigSerializer(serializers.ModelSerializer):
 
     data = serializers.JSONField(source="content")  # type: ignore
     key = serializers.CharField(source="slug")
-
-
-def is_round_over(round):
-    if not round.ended_at:
-        return False
-    return round.ended_at > date.today()
-
-
-def sort_campaigns_by_last_round_end_date(campaign):
-    sorted_rounds = sorted(
-        list(campaign.rounds.all()),
-        key=lambda round: round.ended_at if round.ended_at else date.max,
-        reverse=True,
-    )
-    return sorted_rounds[0].ended_at if sorted_rounds[0].ended_at else date.max
-
-
-def determine_status_for_district(district_data):
-    if not district_data:
-        return "0"
-    checked = district_data["total_child_checked"]
-    marked = district_data["total_child_fmd"]
-    if checked == 60:
-        if marked > 56:
-            return "1lqasOK"
-        return "3lqasFail"
-    return "2lqasDisqualified"
-
-
-def reduce_to_country_status(total, current):
-    if not total.get("passed", None):
-        total["passed"] = 0
-    if not total.get("total", None):
-        total["total"] = 0
-    if int(current[0]) == 1:
-        total["passed"] = total["passed"] + 1
-    total["total"] = total["total"] + 1
-    return total
-
-
-def calculate_country_status(country_data, roundNumber="latest"):
-    if len(country_data.get("rounds", [])) == 0:
-        # TODO put in an enum
-        return "inScope"
-    data_for_all_rounds = sorted(country_data["rounds"], key=lambda round: round["number"], reverse=True)
-    if roundNumber == "latest":
-        data_for_round = data_for_all_rounds[0]
-    else:
-        data_for_round = next((round for round in data_for_all_rounds if round["number"] == int(roundNumber)), None)
-
-    district_statuses = [
-        determine_status_for_district(district_data) for district_data in data_for_round["data"].values()
-    ]
-    aggregated_statuses = reduce(reduce_to_country_status, district_statuses, {})
-    if aggregated_statuses.get("total", 0) == 0:
-        return "inScope"
-    passing_ratio = round((aggregated_statuses["passed"] * 100) / aggregated_statuses["total"])
-    if passing_ratio >= 80:
-        return "1lqasOK"
-    if passing_ratio >= 50:
-        return "2lqasDisqualified"
-    return "3lqasFail"
-
-
-class LQASIMDataStoreSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = JsonDataStore
-        fields = ["created_at", "updated_at", "key", "data"]
-
-    data = serializers.JSONField(source="content")  # type: ignore
-    key = serializers.CharField(source="slug")
-
-    def to_representation(self, instance):
-        request = self.context["request"]
-        representation = super().to_representation(instance)
-        country_id = int(representation["key"].split("_")[1])
-        org_unit = OrgUnit.objects.get(pk=country_id).as_dict()
-        shapes = None
-        if org_unit["has_geo_json"] == True:
-            shape_queryset = OrgUnit.objects.filter_for_user_and_app_id(
-                request.user, request.query_params.get("app_id", None)
-            ).filter(id=org_unit["id"])
-            shapes = geojson_queryset(shape_queryset, geometry_field="simplified_geom")
-
-        # Probably not necessary as long as we only have AFRO in the platform
-        campaigns = Campaign.objects.filter(country=country_id).filter(deleted_at=None)
-        finished_campaigns = [
-            campaign
-            for campaign in campaigns
-            if len([round for round in campaign.rounds.all() if is_round_over(round)]) == 0
-        ]
-
-        # TODO make this null safe
-        latest_campaign = sorted(
-            finished_campaigns,
-            key=lambda campaign: sort_campaigns_by_last_round_end_date(campaign),
-            reverse=True,
-        )[0]
-
-        # TODO make null safe
-        data_for_country = representation["data"]
-        # remove data from all campaigns but latest
-        stats = data_for_country.get("stats", None)
-        if stats:
-            stats = stats.get(latest_campaign.obr_name, None)
-        if stats:
-            return {
-                "id": int(country_id),
-                "data": {"campaign": latest_campaign.obr_name, **stats},
-                "shapes": shapes,
-                "status": calculate_country_status(stats),
-            }
-        return {
-            "id": int(country_id),
-            "data": {"campaign": latest_campaign.obr_name},
-            "shapes": shapes,
-            "status": calculate_country_status({}),
-        }
